Move `InvalidAssignee` to `UserError`
